Move CListBox to UI
