+++ src/game/client/components/menus.h
@@ -16,6 +16,7 @@
 #include <game/client/component.h>
 #include <game/client/localization.h>
 #include <game/client/ui.h>
+#include <game/client/ui_listbox.h>
 #include <game/client/ui_scrollregion.h>
 
 #include "skins.h"
@@ -54,60 +55,6 @@ private:
 	void DoJoystickBar(const CUIRect *pRect, float Current, float Tolerance, bool Active);
 	void DoButton_KeySelect(CButtonContainer *pBC, const char *pText, const CUIRect *pRect);
 	int DoKeyReader(CButtonContainer *pPC, const CUIRect *pRect, int Key, int Modifier, int *pNewModifier);
-
-	// Listbox : found in menus_listbox.cpp
-	struct CListboxItem
-	{
-		bool m_Visible;
-		bool m_Selected;
-		bool m_Disabled;
-		CUIRect m_Rect;
-	};
-
-	// Instances of CListBox must be static, as member addresses are used as UI item IDs
-	class CListBox : private CUIElementBase
-	{
-	private:
-		CUIRect m_ListBoxView;
-		float m_ListBoxRowHeight;
-		int m_ListBoxItemIndex;
-		int m_ListBoxSelectedIndex;
-		int m_ListBoxNewSelected;
-		int m_ListBoxNewSelOffset;
-		int m_ListBoxUpdateScroll;
-		int m_ListBoxDoneEvents;
-		int m_ListBoxNumItems;
-		int m_ListBoxItemsPerRow;
-		bool m_ListBoxItemActivated;
-		const char *m_pBottomText;
-		float m_FooterHeight;
-		CScrollRegion m_ScrollRegion;
-		vec2 m_ScrollOffset;
-		char m_aFilterString[64];
-		CLineInput m_FilterInput;
-		int m_BackgroundCorners;
-
-	protected:
-		CListboxItem DoNextRow();
-
-	public:
-		CListBox();
-
-		void DoBegin(const CUIRect *pRect);
-		void DoHeader(const CUIRect *pRect, const char *pTitle, float HeaderHeight = 20.0f, float Spacing = 2.0f);
-		void DoSpacing(float Spacing = 20.0f);
-		bool DoFilter(float FilterHeight = 20.0f, float Spacing = 2.0f);
-		void DoFooter(const char *pBottomText, float FooterHeight = 20.0f); // call before DoStart to create a footer
-		void DoStart(float RowHeight, int NumItems, int ItemsPerRow, int RowsPerScroll, int SelectedIndex,
-					const CUIRect *pRect = 0, bool Background = true, bool *pActive = 0, int BackgroundCorners = CUIRect::CORNER_ALL);
-		CListboxItem DoNextItem(const void *pID, bool Selected = false, bool *pActive = 0);
-		CListboxItem DoSubheader();
-		int DoEnd();
-		bool FilterMatches(const char *pNeedle) const;
-		bool WasItemActivated() const { return m_ListBoxItemActivated; }
-		float GetScrollBarWidth() const { return m_ScrollRegion.IsScrollbarShown() ? 20 : 0; } // defined in menus_scrollregion.cpp
-	};
-
 
 	enum
 	{
